requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Script==2.0.6
gunicorn==19.8.1
idna==2.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
newsapi-python==0.2.3
requests==2.17.1
urllib3==1.21.1
visitor==0.1.3
Werkzeug==0.14.1
